+++ packages/@podman-desktop-companion/container-app/src/bridge/setup.js
@@ -1,21 +1,21 @@
 // vendors
 const { createLogger } = require("@podman-desktop-companion/logger");
 // locals
-const logger = createLogger("bridge.dummy");
+const logger = createLogger("bridge.setup");
 
-const dummyFunction = async (currentApi, options) => {
-  logger.debug("Dummy function call");
-  return true;
+const setup = async (currentApi, options) => {
+  logger.debug("Setup");
+  return { logger: createLogger("shell.ui") };
 };
 
 function createActions(context, { ipcRenderer, userConfiguration, osType, version, environment }) {
   // Do not access the context at creation - it is lazy
   return {
-    dummyFunction: (...rest) => dummyFunction(context.getCurrentApi(), ...rest)
+    setup: (...rest) => setup(context.getCurrentApi(), ...rest)
   };
 }
 
 module.exports = {
-  dummyFunction,
+  setup,
   createActions
 };